test/reftest/web-platform-tests/dom_converted/events/Event-dispatch-target-moved.html

<!DOCTYPE html>
<meta charset=utf-8>
<!--EDIT-->
<!--title> Determined event propagation path - target moved </title-->
<!--link rel="help" href="https://dom.spec.whatwg.org/#concept-event-dispatch"-->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #table {
    display: none;
  }
  .col {
    display: inline-block;
  }
  .cell {
    border: 1px solid black;
    padding: 0;
    margin: 1px;
  }
</style>
<div id=log></div>
<div id="table">
  <div id="table-body">
    <div class="col" id="parent">
      <p class="cell">Shady Grove</p>
      <p class="cell" id="target">Over the river, Charlie</p>
    </div>
    <div class="col">
      <p class="cell">Aeolian</p>
      <p class="cell">Dorian</p>
    </div>
    <div class="col" id="table-row"></div>
  </div>
</div>
<script>
test(function() {
    var event_type = "foo";
    var target = document.getElementById("target");
    var parent = document.getElementById("parent");
    var tbody = document.getElementById("table-body");
    var table = document.getElementById("table");
    var body = document.body;
    var html = document.documentElement;
    var targets = [window, document, html, body, table, tbody, parent, target];
    var expected_targets = targets.concat([target, parent, tbody, table, body, html, document, window]);
    var phases = [
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.CAPTURING_PHASE,
        Event.AT_TARGET,
        Event.AT_TARGET,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
        Event.BUBBLING_PHASE,
    ];

    var actual_targets = [], actual_phases = [];
    var test_event = this.step_func(function(evt) {
        if (parent === target.parentNode) {
            var table_row = document.getElementById("table-row");
            table_row.appendChild(parent.removeChild(target));
        }

        actual_targets.push(evt.currentTarget);
        actual_phases.push(evt.eventPhase);
    });

    for (var i = 0; i < targets.length; i++) {
        targets[i].addEventListener(event_type, test_event, true);
        targets[i].addEventListener(event_type, test_event, false);
    }
    // EDIT
    // var evt = document.createEvent("Event");
    // evt.initEvent(event_type, true, true);
    var evt = new Event(event_type, {"bubbles": true, "cancelable": true});
    target.dispatchEvent(evt);

    assert_array_equals(actual_targets, expected_targets, "targets");
    assert_array_equals(actual_phases, phases, "phases");
}, "Event propagation path when an element in it is moved within the DOM");
</script>
